Added File Selector for 2D y-axis
Added the option to select different files when selecting the value to plot on the Y-axis of the 2D graphs. Also, commented out something I thought wont be needed, left it just in case during testing we find out it is needed. It is related to the Create new 2D graph button